API

Index

ScienceProjectTemplate.Stats

ScienceProjectTemplate.cartesian_list

ScienceProjectTemplate.check_filename

ScienceProjectTemplate.combine_results

ScienceProjectTemplate.subseteq

Docs

Stats

ScienceProjectTemplate.Stats — Type

Stats() <: OnlineStat

A type for collecting statistics.

Examples

julia> using OnlineStats

julia> s = Stats();

julia> OnlineStats.fit!(s, (a = 1, b = 2));  # add one datapoint at a time

julia> OnlineStats.fit!(s, Dict(:a => 2, :b => 3, :c=>4)); # also support dicts and new observables

julia> s
Stats:
  a  =  1.5 ± 0.5       (2 obs)
  b  =  2.5 ± 0.5       (2 obs)
  c  =  4.0 ± Inf       (1 obs)

julia> data = [(a = i, b = 2*i) for i in 1:10];

julia> OnlineStats.fit!(s, data);  # add multiple datapoints

julia> OnlineStats.nobs(s)
12

julia> s.a      # mean and error (as a Measurements.jl type)
4.83 ± 0.91

julia> reduce(Stats(), data)
Stats:
  a  =  5.5 ± 0.96      (10 obs)
  b  =  11.0 ± 1.9      (10 obs)

Misc

ScienceProjectTemplate.cartesian_list — Method

cartesian_list(; kws...)

Return a vector containing all of the combinations of the values in the keyword arguments. Similar to collect(Iterators.product(kws...)) and to DrWatson.dict_list but returns a vector of NamedTuples.

Examples

julia> cartesian_list(a = [1,2], b = [3,4])
2-element Vector{NamedTuple{(:a, :b),Tuple{Int64,Int64}}}:
 (a = 1, b = 3)
 (a = 1, b = 4)
 (a = 2, b = 3)
 (a = 2, b = 4)

ScienceProjectTemplate.check_filename — Method

check_filename(filename)

Check if a file with the name filename exists, if so, append a number to the name.

ScienceProjectTemplate.combine_results — Method

combine_results(df; by, cols, errs = nothing, col_n = :nsamples)

Combine the results of measurements in df by averaging the values in cols grouped by the columns in by and propagating the errors in errs using the number of samples in col_n.

If no errors are given, the standard deviation on the mean is used as the error.

The final error is given by sqrt(var / (n*(n-1))) where var is the variance and n is the total number of samples.

Examples

julia> using DataFrames

julia> df = DataFrame(a = [1,1,3], b=[2,2,4], n = [2,3,4], c = [2.,2.1,4.], c_err = [0.2,0.1,0.2])
3×5 DataFrame
 Row │ a      b      n      c        c_err   
     │ Int64  Int64  Int64  Float64  Float64 
─────┼───────────────────────────────────────
   1 │     1      2      2      2.0      0.2
   2 │     1      2      3      2.1      0.1
   3 │     3      4      4      4.0      0.2

julia> combine_results(df, by=1:2, cols=4:2:ncol(df), errs=5:2:ncol(df), col_n=:n)
2×5 DataFrame
 Row │ a      b      n      c        c_err    
     │ Int64  Int64  Int64  Float64  Float64  
─────┼────────────────────────────────────────
   1 │     1      2      5     2.06  0.087178
   2 │     3      4      4     4.0   0.2

julia> df = DataFrame(a=[1,1,3,1], b=[1,2,6,3], c=[7,8,9,10])
4×3 DataFrame
Row │ a      b      c     
    │ Int64  Int64  Int64 
─────┼─────────────────────
    1 │     1      1      7
    2 │     1      2      8
    3 │     3      6      9
    4 │     1      3     10

julia> combine_results(df, by=:a, cols=:b)
2×4 DataFrame
Row │ a      nsamples  b        b_err     
    │ Int64  Int64     Float64  Float64   
─────┼─────────────────────────────────────
    1 │     1         3      2.0    0.57735
    2 │     3         1      6.0  Inf

ScienceProjectTemplate.subseteq — Method

subseteq(df; col1=val1, col2=val2, ...)

Return a subset of df where the values of the columns col1, col2, ..., are equal to val1, val2, ... .

Examples

julia> df = DataFrame(a=[1,1,3,1], b=[4,5,6,4], c=[7,8,9,10])
4×3 DataFrame
 Row │ a      b      c     
     │ Int64  Int64  Int64 
─────┼─────────────────────
   1 │     1      4      7
   2 │     1      5      8
   3 │     3      6      9
   4 │     1      4     10

julia> subseteq(df, a=1, b=4)
2×3 DataFrame
 Row │ a      b      c     
     │ Int64  Int64  Int64 
─────┼─────────────────────
   1 │     1      4      7
   2 │     1      4     10